refactor: add a new ServiceCatalog object
It is currently not publicly accessible, but will replace async
catalog operations later on.
